+++ resources/views/datatable.blade.php
@@ -9,7 +9,7 @@
         <x-livewire-tables::tools>
             <x-livewire-tables::tools.sorting-pills />
             <x-livewire-tables::tools.filter-pills />
-            <x-livewire-tables::tools.toolbar />
+            <x-livewire-tables::tools.toolbar :$filterGenericData />
         </x-livewire-tables::tools>
 
         <x-livewire-tables::table>
@@ -28,7 +28,7 @@
             </x-slot>
 
             @if($this->secondaryHeaderIsEnabled() && $this->hasColumnsWithSecondaryHeader())
-                <x-livewire-tables::table.tr.secondary-header :rows="$rows" />
+                <x-livewire-tables::table.tr.secondary-header :rows="$rows" :$filterGenericData />
             @endif
             @if($this->hasDisplayLoadingPlaceholder())
                 <x-livewire-tables::includes.loading colCount="{{ $this->columns->count()+1 }}" />
@@ -62,9 +62,9 @@
             @if ($this->footerIsEnabled() && $this->hasColumnsWithFooter())
                 <x-slot name="tfoot">
                     @if ($this->useHeaderAsFooterIsEnabled())
-                        <x-livewire-tables::table.tr.secondary-header :rows="$rows" />
+                        <x-livewire-tables::table.tr.secondary-header :rows="$rows" :$filterGenericData />
                     @else
-                        <x-livewire-tables::table.tr.footer :rows="$rows" />
+                        <x-livewire-tables::table.tr.footer :rows="$rows"  :$filterGenericData />
                     @endif
                 </x-slot>
             @endif