+++ src/Views/Traits/Configuration/FilterConfiguration.php
@@ -125,4 +125,18 @@ trait FilterConfiguration
 
         return $this;
     }
+
+    public function setGenericDisplayData(array $genericDisplayData = []): self
+    {
+        $this->genericDisplayData = [
+            'filterLayout' => $genericDisplayData['filterLayout'],
+            'tableName' => $genericDisplayData['tableName'],
+            'isTailwind' => $genericDisplayData['isTailwind'],
+            'isBootstrap' => ($genericDisplayData['isBootstrap4'] || $genericDisplayData['isBootstrap5']),
+            'isBootstrap4' => $genericDisplayData['isBootstrap4'],
+            'isBootstrap5' => $genericDisplayData['isBootstrap5'],
+        ];
+
+        return $this;
+    }
 }