+++ src/Views/Traits/Helpers/ColumnHelpers.php
@@ -341,7 +341,7 @@ trait ColumnHelpers
      * @param  mixed  $rows
      * @return mixed
      */
-    public function getSecondaryHeaderContents($rows)
+    public function getSecondaryHeaderContents($rows, array $filterGenericData)
     {
         $value = null;
         $callback = $this->getSecondaryHeaderCallback();
@@ -353,12 +353,12 @@ trait ColumnHelpers
                     return new HtmlString($value);
                 }
             } elseif ($callback instanceof Filter) {
-                return $callback->setFilterPosition('header')->render($this->getComponent()->filterLayout, $this->getComponent()->getTableName(), $this->getComponent()->isTailwind(), $this->getComponent()->isBootstrap4(), $this->getComponent()->isBootstrap5());
+                return $callback->setFilterPosition('header')->setFilterGenericData($filterGenericData)->render($filterGenericData);
             } elseif (is_string($callback)) {
                 $filter = $this->getComponent()->getFilterByKey($callback);
 
                 if ($filter instanceof Filter) {
-                    return $filter->setFilterPosition('header')->render($this->getComponent()->filterLayout, $this->getComponent()->getTableName(), $this->getComponent()->isTailwind(), $this->getComponent()->isBootstrap4(), $this->getComponent()->isBootstrap5());
+                    return $filter->setFilterPosition('header')->setFilterGenericData($filterGenericData)->render($filterGenericData);
                 }
             } else {
                 throw new DataTableConfigurationException('The secondary header callback must be a closure, filter object, or filter key if using secondaryHeaderFilter().');
@@ -390,7 +390,7 @@ trait ColumnHelpers
      * @param  mixed  $rows
      * @return mixed
      */
-    public function getFooterContents($rows)
+    public function getFooterContents($rows, array $filterGenericData)
     {
         $value = null;
         $callback = $this->getFooterCallback();
@@ -403,12 +403,12 @@ trait ColumnHelpers
                     return new HtmlString($value);
                 }
             } elseif ($callback instanceof Filter) {
-                return $callback->setFilterPosition('footer')->render($this->getComponent()->filterLayout, $this->getComponent()->getTableName(), $this->getComponent()->isTailwind(), $this->getComponent()->isBootstrap4(), $this->getComponent()->isBootstrap5());
+                return $callback->setFilterPosition('footer')->setFilterGenericData($filterGenericData)->render($filterGenericData);
             } elseif (is_string($callback)) {
                 $filter = $this->getComponent()->getFilterByKey($callback);
 
                 if ($filter instanceof Filter) {
-                    return $filter->setFilterPosition('footer')->render($this->getComponent()->filterLayout, $this->getComponent()->getTableName(), $this->getComponent()->isTailwind(), $this->getComponent()->isBootstrap4(), $this->getComponent()->isBootstrap5());
+                    return $filter->setFilterPosition('footer')->setFilterGenericData($filterGenericData)->render($filterGenericData);
                 }
             } else {
                 throw new DataTableConfigurationException('The footer callback must be a closure, filter object, or filter key if using footerFilter().');