+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to `laravel-livewire-tables` will be documented in this file
 
+## UNRELEASED
+- Non-Impacting - Modify Filters to use DTO for Generic Properties
+
 ## [3.0.0-beta.11] - 2023-10-29
 - Update Date Range Icon Styling
 - Migrate PHP from date-range blade into new DateRangeFilter method